move core tests - needs timing fix
