+++ ReactWindows/ReactNative.Shared.Tests/Modules/Core/RCTNativeAppEventEmitterTests.cs
@@ -1,14 +1,15 @@
-﻿using Microsoft.VisualStudio.TestPlatform.UnitTestFramework;
+﻿using NUnit.Framework;
 using ReactNative.Modules.Core;
 
 namespace ReactNative.Tests.Modules.Core
 {
-    [TestClass]
-    public class RCTDeviceEventEmitterTests
+    [TestFixture]
+    public class RCTNativeAppEventEmitterTests
     {
-        public void RCTDeviceEventEmitter_Invoke()
+        [Test]
+        public void RCTEventEmitter_InvokeTests()
         {
-            var module = new RCTDeviceEventEmitter();
+            var module = new RCTNativeAppEventEmitter();
 
             var name = default(string);
             var args = default(object[]);
@@ -21,7 +22,7 @@ namespace ReactNative.Tests.Modules.Core
             var eventName = "foo";
             var data = new object();
             module.emit(eventName, data);
-            Assert.AreEqual(nameof(RCTDeviceEventEmitter.emit), name);
+            Assert.AreEqual(nameof(RCTNativeAppEventEmitter.emit), name);
             Assert.AreEqual(2, args.Length);
             Assert.AreSame(eventName, args[0]);
             Assert.AreSame(data, args[1]);
