+++ ReactWindows/ReactNative.Shared.Tests/Modules/Core/TimingTests.cs
@@ -1,4 +1,4 @@
-﻿using Microsoft.VisualStudio.TestPlatform.UnitTestFramework;
+﻿using NUnit.Framework;
 using ReactNative.Bridge;
 using ReactNative.Modules.Core;
 using System;
@@ -6,14 +6,13 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
-using Windows.System.Threading;
 
 namespace ReactNative.Tests.Modules.Core
 {
-    [TestClass]
+    [TestFixture]
     public class TimingTests
     {
-        [TestMethod]
+        [Test]
         public async Task Timing_Create()
         {
             var ids = new List<int>();
@@ -37,7 +36,7 @@ namespace ReactNative.Tests.Modules.Core
             await DispatcherHelpers.CallOnDispatcherAsync(context.DisposeAsync);
         }
 
-        [TestMethod]
+        [Test]
         public async Task Timing_ManyTimers()
         {
             var count = 1000;
@@ -69,7 +68,7 @@ namespace ReactNative.Tests.Modules.Core
             await DispatcherHelpers.CallOnDispatcherAsync(context.DisposeAsync);
         }
 
-        [TestMethod]
+        [Test]
         public async Task Timing_Create_Delete()
         {
             var ids = new List<int>();
@@ -94,7 +93,7 @@ namespace ReactNative.Tests.Modules.Core
             await DispatcherHelpers.CallOnDispatcherAsync(context.DisposeAsync);
         }
 
-        [TestMethod]
+        [Test]
         public async Task Timing_Suspend_Resume()
         {
             var ids = new List<int>();
@@ -120,7 +119,7 @@ namespace ReactNative.Tests.Modules.Core
             await DispatcherHelpers.CallOnDispatcherAsync(context.DisposeAsync);
         }
 
-        [TestMethod]
+        [Test]
         public async Task Timing_Repeat()
         {
             var ids = new List<int>();
@@ -133,7 +132,8 @@ namespace ReactNative.Tests.Modules.Core
                 ids.AddRange((IList<int>)args[0]);
                 if (countdown.CurrentCount > 0)
                 {
-                    var t = ThreadPool.RunAsync(_ => countdown.Signal());
+                    Action runCode = () => { countdown.Signal(); };
+                    Task.Run(runCode).ConfigureAwait(false);
                 }
             }));
 
@@ -155,7 +155,7 @@ namespace ReactNative.Tests.Modules.Core
             await DispatcherHelpers.CallOnDispatcherAsync(context.DisposeAsync);
         }
 
-        [TestMethod]
+        [Test]
         public async Task Timing_ManOrBoy()
         {
             var r = new Random();
@@ -198,7 +198,7 @@ namespace ReactNative.Tests.Modules.Core
             await DispatcherHelpers.CallOnDispatcherAsync(context.DisposeAsync);
         }
 
-        [TestMethod]
+        [Test]
         public async Task Timing_AnimationBehavior()
         {
             var id = 0;
@@ -248,7 +248,7 @@ namespace ReactNative.Tests.Modules.Core
             await DispatcherHelpers.CallOnDispatcherAsync(context.DisposeAsync);
         }
 
-        [TestMethod]
+        [Test]
         public async Task Timing_Zero_NoRepeat()
         {
             var ids = new List<int>();
@@ -272,7 +272,7 @@ namespace ReactNative.Tests.Modules.Core
             await DispatcherHelpers.CallOnDispatcherAsync(context.DisposeAsync);
         }
 
-        [TestMethod]
+        [Test]
         public async Task Timing_Zero_Repeat()
         {
             var ids = new List<int>();
@@ -296,7 +296,7 @@ namespace ReactNative.Tests.Modules.Core
             await DispatcherHelpers.CallOnDispatcherAsync(context.DisposeAsync);
         }
 
-        [TestMethod]
+        [Test]
         public async Task Timing_Correct_Order()
         {
             var ids = new List<int>();
